some changes towards more OOP
